refactor: TOP-16 picture -> profileImage 로 이름 변경
